Create container and call API
