+++ frontend/src/App.js
@@ -2,14 +2,14 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 import MomentUtils from "@date-io/moment";
 import { MuiPickersUtilsProvider } from "@material-ui/pickers";
 
-import Registration from "./Registration";
+import RegistrationContainer from "./containers/RegistrationContainer/RegistrationContainer";
 
 function App() {
 	return (
 		<>
 			<CssBaseline />
 			<MuiPickersUtilsProvider utils={MomentUtils}>
-				<Registration />
+				<RegistrationContainer />
 			</MuiPickersUtilsProvider>
 		</>
 	);
