+++ frontend/src/properties/endpoints.js
@@ -1,0 +1,3 @@
+const ACCOUNT = "http://localhost:8080/api/v1/account";
+
+export { ACCOUNT };
